Updated environment variables in README
# Overview
I added / changed the environment variable `AIRFLOW_WEBSERVER_HOST=airflow`  for the airflow worker / scheduler services in the docker cli instructions to resolve an issue with the current version. In the current version, one is unable to run any DAGS due to a known issue with that environment variable being incorrect. The issue one would get is:  `timeout reached before the port went into state "inuse"`.

## Related issues
https://github.com/bitnami/bitnami-docker-airflow/issues/78
https://github.com/bitnami/bitnami-docker-airflow/issues/87